refact(mutators) rename LCR and SBR
