+++ cas9-engine/src/main/java/org/pitest/mutationtest/engine/cas9/config/Cas9EngineConfiguration.java
@@ -12,6 +12,7 @@ import lombok.experimental.Accessors;
 import lombok.val;
 import org.pitest.functional.prelude.Prelude;
 import org.pitest.mutationtest.EngineArguments;
+import org.pitest.mutationtest.engine.cas9.mutators.Cas9Mutators;
 import org.pitest.mutationtest.engine.gregor.MethodInfo;
 import org.pitest.mutationtest.engine.gregor.MethodMutatorFactory;
 import org.pitest.mutationtest.engine.gregor.MutationEngineConfiguration;
